Revent renaming of src\lib\data-fetcher.ts
