+++ samples/nextjs/src/lib/data-fetcher.ts
